phase-1-fundamentals/02-smolagents-deep-dive/requirements.txt
smolagents[litellm,toolkit]
python-dotenv
gradio
langchain
langchain-community
